+++ src/Entity/Annotation/ManyToMany.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Newage\Annotations\Entity\Annotation;
+
+/**
+ * @Annotation
+ */
+class ManyToMany extends AbstractAnnotation
+{
+    /**
+     * Retrieve the name
+     *
+     * @return null|string
+     */
+    public function getName()
+    {
+        return $this->value;
+    }
+}