+++ src/Mapper/Annotation/PropertyAnnotationListener.php
@@ -2,9 +2,7 @@
 
 namespace Newage\Annotations\Mapper\Annotation;
 
-use Newage\Annotations\Entity\Annotation\Column;
-use Newage\Annotations\Entity\Annotation\Id;
-use Newage\Annotations\Entity\Annotation\OneToOne;
+use Newage\Annotations\Entity\Annotation;
 use Zend\EventManager\EventInterface;
 use Zend\EventManager\EventManagerInterface;
 
@@ -19,12 +17,14 @@ class PropertyAnnotationListener extends AbstractAnnotationListener
         $this->listeners[] = $events->attach('configureProperty', [$this, 'handleIdAnnotation']);
         $this->listeners[] = $events->attach('configureProperty', [$this, 'handleColumnAnnotation']);
         $this->listeners[] = $events->attach('configureProperty', [$this, 'handleOneToOneAnnotation']);
+        $this->listeners[] = $events->attach('configureProperty', [$this, 'handleOneToManyAnnotation']);
+        $this->listeners[] = $events->attach('configureProperty', [$this, 'handleManyToManyAnnotation']);
     }
 
     public function handleIdAnnotation(EventInterface $event)
     {
         $annotation = $event->getParam('annotation');
-        if (!$annotation instanceof Id) {
+        if (!$annotation instanceof Annotation\Id) {
             return;
         }
 
@@ -35,7 +35,7 @@ class PropertyAnnotationListener extends AbstractAnnotationListener
     public function handleColumnAnnotation(EventInterface $event)
     {
         $annotation = $event->getParam('annotation');
-        if (!$annotation instanceof Column) {
+        if (!$annotation instanceof Annotation\Column) {
             return;
         }
 
@@ -46,11 +46,33 @@ class PropertyAnnotationListener extends AbstractAnnotationListener
     public function handleOneToOneAnnotation(EventInterface $event)
     {
         $annotation = $event->getParam('annotation');
-        if (!$annotation instanceof OneToOne) {
+        if (!$annotation instanceof Annotation\OneToOne) {
             return;
         }
 
         $spec = $event->getParam('spec');
         $spec['oneToOne'] = $annotation->getName();
+    }
+
+    public function handleOneToManyAnnotation(EventInterface $event)
+    {
+        $annotation = $event->getParam('annotation');
+        if (!$annotation instanceof Annotation\OneToMany) {
+            return;
+        }
+
+        $spec = $event->getParam('spec');
+        $spec['oneToMany'] = $annotation->getName();
+    }
+
+    public function handleManyToManyAnnotation(EventInterface $event)
+    {
+        $annotation = $event->getParam('annotation');
+        if (!$annotation instanceof Annotation\ManyToMany) {
+            return;
+        }
+
+        $spec = $event->getParam('spec');
+        $spec['manyToMany'] = $annotation->getName();
     }
 }